Implement TranslatorInteractor.kt as a glue between VM&Repository
