+++ repository/build.gradle
@@ -47,4 +47,7 @@ dependencies {
     api 'com.jakewharton.retrofit:retrofit2-kotlinx-serialization-converter:0.8.0'
     api 'com.squareup.okhttp3:logging-interceptor:4.9.0'
     implementation "org.jetbrains.kotlinx:kotlinx-serialization-json:1.0.1"
+
+    //Coroutines
+    implementation 'org.jetbrains.kotlinx:kotlinx-coroutines-android:1.4.1'
 }
